add policy evaluation tests to task stage integration tests
